requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-google-genai
langchain-openai
langchain-groq
langchain-community
langchain-qdrant
qdrant-client
fastapi-mail
slowapi
python-multipart
requests
alembic
